test(serialization): dispatch remaining tests
