+++ test/model/test_quantize_mlp.py
@@ -1,8 +1,11 @@
+import io
+
 import pytest
 import torch
-from helpers import assert_similar, random_qtensor
+from helpers import assert_similar, random_qtensor, random_tensor
 
 from quanto import Calibration, QLinear, QTensor, freeze, qfloat8_e4m3fn, qfloat8_e5m2, qint8, quantize
+from quanto.nn import QModuleMixin
 
 
 class MLP(torch.nn.Module):
@@ -71,3 +74,34 @@ def test_quantize_mlp_int8_activations(weights, frozen, device):
 @pytest.mark.skip_device("mps")
 def test_quantize_mlp_float8_activations(weights, activations, frozen, device):
     _test_quantize_mlp(weights, activations, frozen, device)
+
+
+@pytest.mark.parametrize("weights", [qint8], ids=["w-qint8"])
+@pytest.mark.parametrize("dtype", [torch.float16, torch.float32], ids=["fp16", "fp32"])
+def test_serialize_quantized_mlp(weights, dtype, device):
+    if dtype == torch.float16 and device.type == "cpu":
+        pytest.skip("Matrix multiplication is not supported for float16 on CPU")
+    input_features = 32
+    hidden_features = 10
+    output_features = 128
+    model = MLP(input_features, hidden_features, output_features).to(dtype).to(device)
+    quantize(model, weights=weights)
+    inputs = random_tensor((1, 10, input_features), dtype=dtype).to(device)
+    with Calibration():
+        model(inputs)
+    freeze(model)
+    b = io.BytesIO()
+    torch.save(model.state_dict(), b)
+    b.seek(0)
+    state_dict = torch.load(b)
+    model_reloaded = MLP(input_features, hidden_features, output_features).to(device)
+    quantize(model_reloaded)
+    # When reloading we must assign instead of copying to force quantized tensors assignment
+    model_reloaded.load_state_dict(state_dict, assign=True)
+    for name, module in model.named_modules():
+        if isinstance(module, QModuleMixin):
+            module_reloaded = getattr(model_reloaded, name)
+            assert torch.equal(module_reloaded.weight._data, module.weight._data)
+            assert torch.equal(module_reloaded.weight._scale, module.weight._scale)
+            assert torch.equal(module_reloaded.input_scale, module.input_scale)
+            assert torch.equal(module_reloaded.output_scale, module.output_scale)
